% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projection_external_transcriptome_dataset.R
\name{projection2seurat}
\alias{projection2seurat}
\title{Projecting external dataset (transcriptome data) to my own dataset}
\arguments{
\item{seurat.obj}{RNAseq.query_dataset}
}
\value{
proj.pca.location: Projected coordinates of public/external dataset on the dimension reduction plot from your own dataset
}
\description{
Projection the public/external dataset using KernelKNN (with pearson correlation option) method to your own dataset
}
\details{
Projection of external dataset

RNAseq.query_dataset: external dataset, seurat.obj: my own dataset
}
