% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat2CellofOrigin.R
\name{seurat2CellofOrigin}
\alias{seurat2CellofOrigin}
\title{cell-of-origin prediction and summary}
\arguments{
\item{seurat.obj}{scHCL.coef.threshold scHCL_prediction_geneset Target.tissue}
}
\value{
scHCL_summary
}
\description{
cell-of-origin using scHCL package
}
\details{
Pipeline for predict cell-of-origin
}
