% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat2dorothea.R
\name{seurat2dorothea}
\alias{seurat2dorothea}
\title{TF-Regulon analysis using DoRothEA package}
\arguments{
\item{seurat.obj}{regulon}
}
\value{
summarized_viper_scores_df
}
\description{
Estimation of gene regulatory network (TF ~ target genes)
}
\details{
Pipeline for regulon analysis
}
