% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integrative_clustering.R
\name{seurat2harmony}
\alias{seurat2harmony}
\title{Integrative clustering}
\arguments{
\item{orig.seurat}{nHVGs dims harmony.dims batch.term}
}
\value{
harmony.seurat
}
\description{
Clustering and Dimension reduction (Seurat) after batch correction with harmony
}
\details{
Integrative clustering.R

TBA
}
