% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAseq_PAM50_prediction.R
\name{seurat2PAM50}
\alias{seurat2PAM50}
\title{PAM50 prediction with Seurat object}
\value{
PAM50xClusters
}
\description{
PAM50 classification with single-cell RNA-seq data
}
\details{
PAM50 prediction

TBA
}
