% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudobulk_DEG_analysis_in_scRNAseq.R
\name{seurat2pseudobulk_DEA}
\alias{seurat2pseudobulk_DEA}
\title{Pseudobulk DEG analysis with Seurat object}
\arguments{
\item{seurat.obj}{metadata}
}
\value{
res_tbl # Data frame of DEG results using DESeq2 package
}
\description{
Aggregation of single-cell RNA-seq data by batch (or specimen)
}
\details{
Pseudobulk differentially expressed gene (DEG) analysis

TBA
}
